requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
pdfplumber>=0.10.3  # Modern PDF text extraction library
pdf2image>=1.17.0  # For converting PDF pages to images
Pillow>=10.0.0
aiohttp>=3.9.0
tqdm>=4.65.0  # For progress bars

# Development dependencies
pytest>=7.4.0
black>=23.9.0
mypy>=1.5.0
pylint>=3.0.0
